requirements/development.txt
Django==1.3.1
psycopg2==2.4.5
PIL==1.1.7
South==0.7.3
Werkzeug==0.6.2
distribute==0.6.10
django-extensions==0.6
django-photologue==2.3
django-rosetta==0.5.6
django-pagination==1.0.7
django-ajax-selects==1.2.5
wsgiref==0.1.2
yolk==0.4.1
